# Better field type detection, fixes #250