<!--
Thank you for proposing a pull request. This template will guide you through the essential steps necessary for a pull request.
-->
Make sure that:

- [x] You have read the [contribution guidelines](https://github.com/mp911de/logstash-gelf/blob/master/.github/CONTRIBUTING.md).
- [ ] You use the code formatters provided [here](https://github.com/mp911de/logstash-gelf/blob/master/as7formatter.xml) and have them applied to your changes. Don’t submit any formatting related changes.
- [x] You submit test cases (unit or integration tests) that back your changes.

 <!--
Great! Live long and prosper.
-->

Performance is slightly worse, especially for Doubles.

Current code:
```
GelfMessageBenchmark.discoverDoubleField                     avgt    5     34,866 ±   4,471  ns/op
GelfMessageBenchmark.discoverLongField                       avgt    5     15,418 ±   2,406  ns/op
GelfMessageBenchmark.discoverStringField                     avgt    5      5,405 ±   1,440  ns/op
```

New code:
```
GelfMessageBenchmark.discoverDoubleField  avgt    5  235,740 ± 5,143  ns/op
GelfMessageBenchmark.discoverLongField    avgt    5   26,484 ± 1,803  ns/op
GelfMessageBenchmark.discoverStringField  avgt    5    5,242 ± 0,254  ns/op
```